requirements.txt
Flask~=3.0
bootstrap-flask~=2.3
Flask-Minify~=0.4
SQLAlchemy~=2.0
Flask-SQLAlchemy~=3.1
alembic~=1.13
flask-login~=0.6
wtforms[email]~=3.1
Flask-WTF~=1.2
pytz~=2024.1
Werkzeug~=3.0
flask-mailman~=1.0
PyJWT~=2.8
